修复(安全): EXIF 日期解析的偏移边界检查改用 64 位运算,不再被回绕绕过
IFD/值的偏移完全来自文件内容。quint32 偏移加法先回绕再比较(如
0xFFFFFFFE+2=0),边界检查恒通过,随后按 ~4GB 的野地址读内存,浏览
含这种构造头的目录即崩。三处比较全部先提升成 quint64 再相加。

+++ src/exifdate.h
@@ -26,9 +26,11 @@ inline double parseTiff(const uchar* p, qint64 avail) {
     quint32 offsets[2] = { ifd0, 0 };
     for (int round = 0; round < 2 && offsets[round]; ++round) {
         quint32 off = offsets[round];
-        if (off + 2 > (quint64)avail) break;
+        // 偏移完全由文件内容控制,先提升成 quint64 再做加法:quint32 加法回绕
+        // (如 off=0xFFFFFFFE 时 off+2=0)会让下面的边界检查形同虚设,直接野读
+        if (quint64(off) + 2 > quint64(avail)) break;
         const quint16 count = u16(p + off);
-        if (off + 2 + count * 12ull > (quint64)avail) break;
+        if (quint64(off) + 2 + count * 12ull > quint64(avail)) break;
         for (quint16 i = 0; i < count; ++i) {
             const uchar* e = p + off + 2 + i * 12;
             const quint16 tag = u16(e);
@@ -42,7 +44,7 @@ inline double parseTiff(const uchar* p, qint64 avail) {
             if (tag == 0x9003 || tag == 0x9004 || tag == 0x0132) {
                 if (type != 2 || cnt < 19) continue;
                 const quint32 valOff = u32(e + 8);   // 20 字节 ASCII 恒外置
-                if (valOff + 19 > (quint64)avail) continue;
+                if (quint64(valOff) + 19 > quint64(avail)) continue;
                 QDateTime dt = QDateTime::fromString(
                     QString::fromLatin1((const char*)p + valOff, 19),
                     "yyyy:MM:dd HH:mm:ss");
